Mhmmm homegrown crypto.. The best kind.
Instead of storing the challenges in a database and having to deal with
race conditions, locks and whatnot, let's use crypto to make the client
pay for it!

The new scheme encodes and encrypts the necessary state (like gen time)
into the "salt", and since we encrypt it, then we know if it originated
from us.